requirements.txt
fastapi>=0.109
uvicorn>=0.27
sqlalchemy>=2.0
pydantic>=2.5
python-multipart>=0.0.6
pywebview>=4.4
faster-whisper>=1.0
aiofiles>=23.2
